refactor(runtime): extract filesystem and apply_patch tools into fs module
The four user-facing fs tools (file_read, file_write, file_list,
apply_patch) move into tool_runner/fs.rs together with every helper
they share:

- resolve_file_path_ext: the workspace-sandbox path resolver.
- named_ws_prefixes / _writable / _readonly: per-agent named-workspace
  allowlist accessors used by both the dispatcher and apply_patch's
  #3662 defense-in-depth check.
- check_absolute_path_inside_workspace: the ACP-path-jail guard plus
  its full `path_check_tests` regression block (#3313, dot-dot-traversal
  fix included).
- maybe_snapshot: the spawn_blocking-wrapped checkpoint snapshot helper
  invoked by file_write and apply_patch before they mutate the
  workspace.
- maybe_dedup_file_read: the #4971 dedup shim that gates file_read on
  `[context_engine] deduplicate_file_reads`.

All helpers are promoted to `pub(super)` so the dispatcher at the
top of mod.rs continues to call them under their original names; no
call site changed. The now-unused `PathBuf` import on mod.rs goes
away in the same hunk (mod.rs still pulls in `Path` for the
ToolExecContext fields).

mod.rs: 10,461 -> 10,043 LOC (-418). cargo check + cargo clippy --lib
both clean.

Refs #3710
